Use webpack to deliver newest ACE editor
With this commit, we change the delivery method for the ACE editor from the `ace-rails-ap` gem to using yarn and webpack. As a side-change, we also modify how the mode is selected through JavaScript instead of Ruby.

Through webpack, the `basePath` is automatically determined and working as expected without any manual intervention.

Besides refactoring the integration of ACE, this commit further resolves an issue where some files could not be fetched correctly in our production environment.