Fix flaky Label test failing due to slow select2 search
I assume that the test is failing because the "Enter" key press does not select the label. During my reproduction setup, this only occurred when the AJAX search was slow, not returning any result in time. As long as there is no result, select2 won't display the option to create a new label. Thus, we could not select it, sending an empty array back to the server. Without any label sent, the error page will not contain any options, thus the test will fail.

Further, we ensure that an element is correctly selected and clicked on, asserting that too.

Note: On the error page, select2 is not necessarily required, and the test can pass when it's not yet initialized.