Add functionality to expand the markdown editor
The editor will now have a default height of 300px but a button will let
the user expand the editor. It will expand it to fit all content (or up
to 400px if the content was not exceeding 300px). In the expanded mode
the editor will keep growing as the user types more content.

Co-authored-by: Julia Casamitjana <62883011+juliacasamitjana@users.noreply.github.com>