Fix incorrect indention for status group icon
When viewing the page in development mode, the file names of views are annotated. Since the icon is extracted into an own partial, it is followed by such a comment. This caused a correct indention, which was missing in production.

The relevant setting is `config.action_view.annotate_rendered_view_with_filenames`.