Disallow edits to the UUID
Currently, sending the UUID through the browser triggers an "Unpermitted Parameter" (for create and update). Hence, we just disallow the option for now, since there are potentially issues with custom UUIDs (especially regarding task contributions). It would be great to allow a manual UUID for new elements (this shouldn't be a problem) or an edit, but both is currently out-of-scope.