Add production-ready Docker image
The docker image is not meant for local development, but for a production-level deployment. It is based on the Docker setup added by default for new Rails apps.

As part of this change, we had to rethink our integration of JavaScript and CSS compressors. Without the change in `production.rb`, an ExecJS runtime would be needed to run the Rails app. Since we require a JavaScript runtime only for the asset compilation (but not when serving the application), this would bloat the Docker image unnecessarily.