Fix TransferValues for recently changed files association
Since we changed the file association in 9d9ac231, the `files` method might return an array of files. For transferring values, however, we need a ActiveRecord::CollectionProxy to perform actual changes on the records. With this commit, we add an exception for the transfer of values to fix that behavior.