feat(rspec): Set session variables during request specs
Since RSpec 3.5, controller specs are deprecated. The official
recommendation of the Rails team and the RSpec core team is to write
request specs instead. They involve the router, the middleware stack,
and both rack requests and responses. Thus, it's not possible to set the
session variables beforehand anymore. Instead, a request spec should call
the sign in endpoint before calling the actual endpoint under test, when
the session is needed.

To avoid the complexity of SSO and SLOs during request tests, this helper
introduces the option to set the session variables via a designated
endpoint for tests.
https://gist.github.com/dteoh/99721c0321ccd18286894a962b5ce584?permalink_comment_id=4188995#gistcomment-4188995