Avoid unstable SQL sort in ProFormA service specs
The existing test relied on the order given by convert_to_task_service.test to match the order given through the tests variable (filled with ProformaXML::Test). However, the ordering is not guaranteed through PostgreSQL and could change, e.g. through an update operation. Hence, it is safer to order the elements first and then perform the expectations.