Allow a list of SENTRY_JAVASCRIPT_DSN
We are preparing CodeHarbor to run under multiple domains (such as open.hpi.de and openhpi.de). To ensure that Sentry is not blocked as a third-party context in any setting, we allow to set multiple DSNs. This change is backward compatible and in line with a similar change in CodeOcean.