Fix button_to for forms
Forms shouldn't be nested, particularly their submit buttons shouldn't. With this commit, we fix this issue. For some forms, we had to provide an ID and assign a separate submit button outside the form to keep the current appearance and make the least changes. A more proper solution should be found in the future.